requirements.txt
pandas
openpyxl
sqlalchemy
psycopg2-binary
python-dotenv
streamlit
scikit-learn
joblib
